Refactor Tests
- Move non tests to testing directory: constants, settings and urls
- Remove coupling between tests nad implementation
- Add fixture scoping for better performance
