+++ pyproject.toml
@@ -51,7 +51,7 @@ build-backend = "setuptools.build_meta"
 packages = ["django_sqlite_backup"]
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "tests.settings"
+DJANGO_SETTINGS_MODULE = "testing.settings"
 
 [tool.black]
 line-length = 99
