+++ testing/constants.py
@@ -1,0 +1,7 @@
+from datetime import datetime
+
+
+TEST_BUCKET_NAME = "mock-bucket"
+TEST_DB_NAME = "fake.db"
+TEST_DATE = datetime(1992, 11, 25)
+TEST_DATE_STR = TEST_DATE.strftime("%Y-%m-%d")
